Coordinate pipes and notif. buffers in the kernel
The kernel now is responsible for queue allocation and deallocation.
Having this centralized in the kernel eliminates the need to for the
hack to figure out pipe IDs based on the core number and number of
queues per core. Pipes can now be allocated in arbitrary order and
different applications (and threads) can use different numbers of pipes.

Fallback queues are also now configured and managed from software. It is
no longer possible to configure those from JTAG.

These changes also required API changes to eliminate the need to specify
number of queues and cores when instantiating a Device as well as to be
able to specify that a given pipe is a fallback pipe.